refactor: 환경 변수변경 및 Components 폴더 위치 수정
