+++ src/pages/StudyListPage/StudyComponents/Card.module.css
