+++ src/pages/StudyListPage/StudyComponents/Search.jsx
