+++ src/pages/StudyListPage/StudyComponents/Search.module.css
