+++ src/pages/StudyListPage/StudyComponents/studyBackground.js
